requirements.txt
amqp==5.1.1
anyio==3.6.2
asgiref==3.6.0
autopep8==2.0.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
billiard==3.6.4.0
black==23.1.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-prometheus==2.2.0
django-silk==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
gprof2dot==2022.7.29
grpcio==1.51.3
grpcio-tools==1.51.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.11.0
pika==1.3.1
platformdirs==2.6.2
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==4.22.1
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7.1
rabbitmq==0.2.0
requests==2.28.2
rfc3986==1.5.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sentry-sdk==1.14.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
sqlparse==0.4.3
structlog==22.3.0
tomli==2.0.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
